WIP: refactoring ClusterStats into cluster_metadata.
